livedns: do not require TTL when specifying a domain
It is not possible currently to query the current TTL, making it
difficult to import and manipulate an existing domain. Make the TTL
optional solves this issue. Also update the documentation: the TTL is
for the SOA, this is not the default one. Also, deprecate it.

Also, set a default value for autosnapshots (matching the current
default value). Otherwise, we have a change each time the resource is
refreshed.